Avoid calculating tile visibility when nothing is changing
