+++ OpenDreamClient/Rendering/DMISpriteSystem.cs
@@ -20,14 +20,18 @@ public sealed class DMISpriteSystem : EntitySystem {
 
     public RenderTargetPool RenderTargetPool = default!;
 
+    private EntityQuery<DMISpriteComponent> _spriteQuery;
     private DreamViewOverlay _mapOverlay = default!;
 
     public override void Initialize() {
         SubscribeLocalEvent<DMISpriteComponent, ComponentAdd>(HandleComponentAdd);
         SubscribeLocalEvent<DMISpriteComponent, ComponentHandleState>(HandleComponentState);
         SubscribeLocalEvent<DMISpriteComponent, ComponentRemove>(HandleComponentRemove);
+        SubscribeLocalEvent<TransformComponent, MoveEvent>(HandleTransformMove);
+        SubscribeLocalEvent<TileChangedEvent>(HandleTileChanged);
 
         RenderTargetPool = new(_clyde);
+        _spriteQuery = _entityManager.GetEntityQuery<DMISpriteComponent>();
         _mapOverlay = new DreamViewOverlay(RenderTargetPool, _transformSystem, _mapSystem, _lookupSystem, _appearanceSystem, _screenOverlaySystem, _clientImagesSystem);
         _overlayManager.AddOverlay(_mapOverlay);
     }
@@ -50,13 +54,26 @@ public sealed class DMISpriteSystem : EntitySystem {
         component.Icon.SizeChanged += () => OnIconSizeChanged(uid);
     }
 
-    private static void HandleComponentState(EntityUid uid, DMISpriteComponent component, ref ComponentHandleState args) {
+    private void HandleComponentState(EntityUid uid, DMISpriteComponent component, ref ComponentHandleState args) {
         SharedDMISpriteComponent.DMISpriteComponentState? state = (SharedDMISpriteComponent.DMISpriteComponentState?)args.Current;
         if (state == null)
             return;
 
+        _mapOverlay.DirtyTileVisibility(); // Our icon's opacity may have changed
         component.ScreenLocation = state.ScreenLocation;
         component.Icon.SetAppearance(state.AppearanceId);
+    }
+
+    private void HandleTransformMove(EntityUid uid, TransformComponent component, ref MoveEvent args) {
+        if (!_spriteQuery.TryGetComponent(uid, out var sprite))
+            return;
+
+        if (sprite.Icon.Appearance?.Opacity is true)
+            _mapOverlay.DirtyTileVisibility(); // A movable with opacity=TRUE has moved
+    }
+
+    private void HandleTileChanged(ref TileChangedEvent ev) {
+        _mapOverlay.DirtyTileVisibility();
     }
 
     private static void HandleComponentRemove(EntityUid uid, DMISpriteComponent component, ref ComponentRemove args) {
