+++ OpenDreamClient/Rendering/DreamViewOverlay.cs
@@ -21,7 +21,7 @@ namespace OpenDreamClient.Rendering;
 /// <summary>
 /// Overlay for rendering world atoms
 /// </summary>
-internal sealed class DreamViewOverlay : Overlay {
+internal sealed partial class DreamViewOverlay : Overlay, IEntityEventSubscriber {
     public static ShaderInstance ColorInstance = default!;
 
     public override OverlaySpace Space => OverlaySpace.WorldSpaceBelowWorld;
@@ -75,9 +75,6 @@ internal sealed class DreamViewOverlay : Overlay {
     private static readonly Matrix3x2 FlipMatrix = Matrix3x2.Identity with {
         M22 = -1
     };
-
-    // Defined here so it isn't recreated every frame
-    private ViewAlgorithm.Tile?[,]? _tileInfo;
 
     public DreamViewOverlay(RenderTargetPool renderTargetPool, TransformSystem transformSystem, MapSystem mapSystem, EntityLookupSystem lookupSystem,
         ClientAppearanceSystem appearanceSystem, ClientScreenOverlaySystem screenOverlaySystem, ClientImagesSystem clientImagesSystem) {
@@ -597,68 +594,6 @@ internal sealed class DreamViewOverlay : Overlay {
                     _planes[planeIndex].DrawMouseMap(handle, this, _mouseMapRenderTarget!.Size, worldAABB);
             }, null);
         }
-    }
-
-    private ViewAlgorithm.Tile?[,] CalculateTileVisibility(EntityUid gridUid, MapGridComponent grid, TileRef eyeTile, int seeVis) {
-        using var _ = _prof.Group("visible turfs");
-
-        var viewRange = _interfaceManager.View;
-        if (_tileInfo == null || _tileInfo.GetLength(0) != viewRange.Width + 2 || _tileInfo.GetLength(1) != viewRange.Height + 2) {
-            // _tileInfo hasn't been created yet or view range has changed, so create a new array.
-            // Leave a 1 tile buffer on each side
-            _tileInfo = new ViewAlgorithm.Tile[viewRange.Width + 2, viewRange.Height + 2];
-        }
-
-        var eyeWorldPos = _mapSystem.GridTileToWorld(gridUid, grid, eyeTile.GridIndices);
-        var tileRefs = _mapSystem.GetTilesEnumerator(gridUid, grid,
-            Box2.CenteredAround(eyeWorldPos.Position, new Vector2(_tileInfo.GetLength(0), _tileInfo.GetLength(1))));
-
-        // Gather up all the data the view algorithm needs
-        while (tileRefs.MoveNext(out var tileRef)) {
-            var delta = tileRef.GridIndices - eyeTile.GridIndices;
-            var appearance = _appearanceSystem.GetTurfIcon((uint)tileRef.Tile.TypeId).Appearance;
-            if (appearance == null)
-                continue;
-
-            int xIndex = delta.X + viewRange.CenterX;
-            int yIndex = delta.Y + viewRange.CenterY;
-            if (xIndex < 0 || yIndex < 0 || xIndex >= _tileInfo.GetLength(0) || yIndex >= _tileInfo.GetLength(1))
-                continue;
-
-            var tile = new ViewAlgorithm.Tile {
-                Opaque = appearance.Opacity,
-                Luminosity = 0,
-                DeltaX = delta.X,
-                DeltaY = delta.Y
-            };
-
-            _tileInfo[xIndex, yIndex] = tile;
-        }
-
-        // Apply entities' opacity
-        foreach (EntityUid entity in EntitiesInView) {
-            // TODO use a sprite tree.
-            if (!_spriteQuery.TryGetComponent(entity, out var sprite))
-                continue;
-
-            var transform = _xformQuery.GetComponent(entity);
-            if (!sprite.IsVisible(transform, seeVis))
-                continue;
-            if (sprite.Icon.Appearance == null) //appearance hasn't loaded yet
-                continue;
-
-            var worldPos = _transformSystem.GetWorldPosition(transform);
-            var tilePos = _mapSystem.WorldToTile(gridUid, grid, worldPos) - eyeTile.GridIndices + viewRange.Center;
-            if (tilePos.X < 0 || tilePos.Y < 0 || tilePos.X >= _tileInfo.GetLength(0) || tilePos.Y >= _tileInfo.GetLength(1))
-                continue;
-
-            var tile = _tileInfo[tilePos.X, tilePos.Y];
-            if (tile != null)
-                tile.Opaque |= sprite.Icon.Appearance.Opacity;
-        }
-
-        ViewAlgorithm.CalculateVisibility(_tileInfo);
-        return _tileInfo;
     }
 
     private void CollectVisibleSprites(ViewAlgorithm.Tile?[,] tiles, EntityUid gridUid, MapGridComponent grid, TileRef eyeTile, sbyte seeVis, SightFlags sight, Box2 worldAABB) {
